+++ RELEASE_NOTES.md
@@ -1,5 +1,14 @@
 # Pushrod Releases
 
+## 0.2.11
+
+- Added image button (#80)
+- Changed Rust Image into a Button
+- Added ability to set button selected state (#89)
+- Modified example application so that it starts the animation on the progress by default
+- Removed debugging from Radio Button.
+- Updated comments in Radio Button.
+
 ## 0.2.10
 
 - Fixed bug with debugging: when selecting checkbox, now toggles hide/show rather than clearing out text. (#116)